requirements.txt
fastapi
uvicorn
redis
pydantic
python-dotenv
email-validator
python-multipart
httpx
slowapi
